requirements.txt
customtkinter
openai
python-dotenv
pygments
